Migrate modules manager and socket tests
