+++ pytests/test_socket.py
@@ -3,6 +3,7 @@
 # for license information.
 
 import pytest
+from ptvsd.socket import Address
 from ptvsd.socket import create_server, close_socket
 
 
@@ -29,3 +30,118 @@ class TestSocketServerReuse(object):
         finally:
             close_socket(sock1)
             close_socket(sock2)
+
+
+class TestAddress(object):
+    def test_from_raw(self):
+        serverlocal = Address.as_server('localhost', 9876)
+        serverremote = Address.as_server('1.2.3.4', 9876)
+        clientlocal = Address.as_client('localhost', 9876)
+        clientremote = Address.as_client('1.2.3.4', 9876)
+        default = Address(None, 1111)
+        external = Address('', 1111)
+        values = [
+            (serverlocal, serverlocal),
+            (serverremote, serverremote),
+            (clientlocal, clientlocal),
+            (clientremote, clientremote),
+            (None, default),
+            ('', external),
+            ([], default),
+            ({}, default),
+            (9876, serverlocal),
+            ('localhost:9876', clientlocal),
+            ('1.2.3.4:9876', clientremote),
+            ('*:9876', Address.as_server('', 9876)),
+            ('*', external),
+            (':9876', Address.as_server('', 9876)),
+            ('localhost', Address('localhost', 1111)),
+            (':', external),
+            (dict(host='localhost'), Address('localhost', 1111)),
+            (dict(port=9876), serverlocal),
+            (dict(host=None, port=9876), serverlocal),
+            (dict(host='localhost', port=9876), clientlocal),
+            (dict(host='localhost', port='9876'), clientlocal),
+        ]
+        for value, expected in values:
+            addr = Address.from_raw(value, defaultport=1111)
+            assert addr == expected
+
+    @pytest.mark.parametrize('host', ['localhost', '127.0.0.1', '::', '1.2.3.4'])
+    def test_as_server_valid_address(self, host):
+        addr = Address.as_server(host, 9786)
+        assert addr == Address(host, 9786, isserver=True)
+
+    def test_as_server_public_host(self):
+        addr = Address.as_server('', 9786)
+        assert addr == Address('', 9786, isserver=True)
+
+    def test_as_server_default_host(self):
+        addr = Address.as_server(None, 9786)
+        assert addr == Address('localhost', 9786, isserver=True)
+
+    @pytest.mark.parametrize('host', [None, '', 'localhost', '1.2.3.4'])
+    def test_as_server_bad_port(self, host):
+        port = None
+        with pytest.raises(TypeError):
+            Address.as_server(host, port)
+
+    @pytest.mark.parametrize('host', [None, '', 'localhost', '1.2.3.4'])
+    @pytest.mark.parametrize('port', ['', -1, 65536])
+    def test_as_server_bad_port2(self, host, port):
+        with pytest.raises(ValueError):
+            Address.as_server(host, port)
+
+    @pytest.mark.parametrize('host', ['localhost', '127.0.0.1', '::', '1.2.3.4'])
+    def test_as_client_valid_address(self, host):
+        addr = Address.as_client(host, 9786)
+        assert addr == Address(host, 9786, isserver=False)
+
+    def test_as_client_public_host(self):
+        addr = Address.as_client('', 9786)
+        assert addr == Address('', 9786, isserver=False)
+
+    def test_as_client_default_host(self):
+        addr = Address.as_client(None, 9786)
+        assert addr == Address('localhost', 9786, isserver=False)
+
+    @pytest.mark.parametrize('host', [None, '', 'localhost', '1.2.3.4'])
+    def test_as_client_bad_port(self, host):
+        port = None
+        with pytest.raises(TypeError):
+            Address.as_client(host, port)
+
+    @pytest.mark.parametrize('host', [None, '', 'localhost', '1.2.3.4'])
+    @pytest.mark.parametrize('port', ['', -1, 65536])
+    def test_as_client_bad_port2(self, host, port):
+        with pytest.raises(ValueError):
+            Address.as_client(host, port)
+
+    @pytest.mark.parametrize('host', ['localhost', '127.0.0.1', '::', '1.2.3.4'])
+    def test_new_valid_address(self, host):
+        addr = Address(host, 9786)
+        assert addr == Address(host, 9786, isserver=False)
+
+    def test_new_public_host(self):
+        addr = Address('', 9786)
+        assert addr == Address('', 9786, isserver=True)
+
+    def test_new_default_host(self):
+        addr = Address(None, 9786)
+        assert addr == Address('localhost', 9786, isserver=True)
+
+    def test_new_wildcard_host(self):
+        addr = Address('*', 9786)
+        assert addr == Address('', 9786, isserver=True)
+
+    @pytest.mark.parametrize('host', [None, '', 'localhost', '1.2.3.4'])
+    def test_new_bad_port(self, host):
+        port = None
+        with pytest.raises(TypeError):
+            Address(host, port)
+
+    @pytest.mark.parametrize('host', [None, '', 'localhost', '1.2.3.4'])
+    @pytest.mark.parametrize('port', ['', -1, 65536])
+    def test_new_bad_port2(self, host, port):
+        with pytest.raises(ValueError):
+            Address(host, port)
